+++ util-stats/src/main/scala/com/twitter/finagle/stats/StatsReceiverProxy.scala
@@ -9,10 +9,9 @@ trait StatsReceiverProxy extends StatsReceiver with DelegatingStatsReceiver {
 
   def repr: AnyRef = self.repr
 
-  def counter(verbosity: Verbosity, names: String*): Counter = self.counter(verbosity, names: _*)
-  def stat(verbosity: Verbosity, names: String*): Stat = self.stat(verbosity, names: _*)
-  def addGauge(verbosity: Verbosity, names: String*)(f: => Float): Gauge =
-    self.addGauge(verbosity, names: _*)(f) // scalafix:ok StoreGaugesAsMemberVariables
+  def counter(schema: CounterSchema): Counter = self.counter(schema)
+  def stat(schema: HistogramSchema): Stat = self.stat(schema)
+  def addGauge(schema: GaugeSchema)(f: => Float): Gauge = self.addGauge(schema)(f)
 
   def underlying: Seq[StatsReceiver] = Seq(self)